+++ lib/oboe/api.rb
@@ -2,6 +2,10 @@
 # All rights reserved.
 
 module Oboe
+  ##
+  # This module implements the TraceView tracing API.
+  # See: https://github.com/appneta/oboe-ruby#the-tracing-api
+  # and/or: http://rdoc.info/gems/oboe/Oboe/API/Tracing
   module API
     def self.extend_with_tracing
       extend Oboe::API::Logging