+++ lib/oboe/base.rb
@@ -22,45 +22,51 @@ SAMPLE_SOURCE_MASK = 0b1111000000000000000000000000
 ZERO_SAMPLE_RATE_MASK   = 0b1111000000000000000000000000
 ZERO_SAMPLE_SOURCE_MASK = 0b0000111111111111111111111111
 
+##
+# This module is the base module for the various implementations of Oboe reporting.
+# Current variations as of 2014-09-10 are a c-extension, JRuby (using TraceView Java
+# instrumentation) and a Heroku c-extension (with embedded tracelyzer)
 module OboeBase
+  extend ::Oboe::ThreadLocal
+
   attr_accessor :reporter
   attr_accessor :loaded
   attr_accessor :sample_source
   attr_accessor :sample_rate
-  attr_accessor :layer_op
+  thread_local :layer_op
 
-  def self.included(cls)
+  def self.included(_)
     self.loaded = true
   end
 
   def tracing_layer_op?(operation)
     if operation.is_a?(Array)
-      return operation.include?(@layer_op)
+      return operation.include?(Oboe.layer_op)
     else
-      return @layer_op == operation
+      return Oboe.layer_op == operation
     end
   end
 
   def always?
-    Oboe::Config[:tracing_mode].to_s == "always"
+    Oboe::Config[:tracing_mode].to_s == 'always'
   end
 
   def never?
-    Oboe::Config[:tracing_mode].to_s == "never"
+    Oboe::Config[:tracing_mode].to_s == 'never'
   end
 
   def passthrough?
-    ["always", "through"].include?(Oboe::Config[:tracing_mode])
+    %w(always through).include?(Oboe::Config[:tracing_mode])
   end
 
   def through?
-    Oboe::Config[:tracing_mode] == "through"
+    Oboe::Config[:tracing_mode] == 'through'
   end
 
   def tracing?
     return false unless Oboe.loaded
 
-    Oboe::Context.isValid and not Oboe.never?
+    Oboe::Context.isValid && !Oboe.never?
   end
 
   def log(layer, label, options = {})
@@ -73,27 +79,30 @@ module OboeBase
   end
 
   def forking_webserver?
-    (defined?(::Unicorn) and ($0 =~ /unicorn/i)) ? true : false
+    (defined?(::Unicorn) && ($PROGRAM_NAME =~ /unicorn/i)) ? true : false
   end
 
   ##
   # These methods should be implemented by the descendants
   # (Oboe_metal, Oboe_metal (JRuby), Heroku_metal)
   #
-  def sample?(opts = {})
-    raise "sample? should be implemented by metal layer."
+  def sample?(_opts = {})
+    fail 'sample? should be implemented by metal layer.'
   end
 
-  def log(layer, label, options = {})
-    raise "log should be implemented by metal layer."
+  def log(_layer, _label, _options = {})
+    fail 'log should be implemented by metal layer.'
   end
 
-  def set_tracing_mode(mode)
-    raise "set_tracing_mode should be implemented by metal layer."
+  def set_tracing_mode(_mode)
+    fail 'set_tracing_mode should be implemented by metal layer.'
   end
 
-  def set_sample_rate(rate)
-    raise "set_sample_rate should be implemented by metal layer."
+  def set_sample_rate(_rate)
+    fail 'set_sample_rate should be implemented by metal layer.'
   end
 end
 
+module Oboe
+  extend OboeBase
+end