+++ lib/oboe/config.rb
@@ -11,10 +11,9 @@ module Oboe
   module Config
     @@config = {}
 
-    @@instrumentation = [ :cassandra, :dalli, :nethttp, :memcached, :memcache, :mongo,
-                          :moped, :rack, :redis, :resque, :action_controller, :action_view,
-                          :active_record, :faraday ]
-
+    @@instrumentation = [:action_controller, :action_view, :active_record,
+                         :cassandra, :dalli, :em_http_request, :faraday, :nethttp, :memcached,
+                         :memcache, :mongo, :moped, :rack, :redis, :resque]
     ##
     # Return the raw nested hash.
     #
@@ -22,7 +21,7 @@ module Oboe
       @@config
     end
 
-    def self.initialize(data={})
+    def self.initialize(_data = {})
       # Setup default instrumentation values
       @@instrumentation.each do |k|
         @@config[k] = {}
@@ -31,6 +30,9 @@ module Oboe
         @@config[k][:log_args] = true
       end
 
+      # Beta instrumentation disabled by default
+      Oboe::Config[:em_http_request][:enabled] = false
+
       # Set collect_backtraces defaults
       Oboe::Config[:action_controller][:collect_backtraces] = true
       Oboe::Config[:active_record][:collect_backtraces] = true
@@ -38,6 +40,7 @@ module Oboe
       Oboe::Config[:cassandra][:collect_backtraces] = true
       Oboe::Config[:dalli][:collect_backtraces] = false
       Oboe::Config[:faraday][:collect_backtraces] = false
+      Oboe::Config[:em_http_request][:collect_backtraces] = false
       Oboe::Config[:memcache][:collect_backtraces] = false
       Oboe::Config[:memcached][:collect_backtraces] = false
       Oboe::Config[:mongo][:collect_backtraces] = true
@@ -59,23 +62,23 @@ module Oboe
       @@config[:blacklist] = []
 
       # Access Key is empty until loaded from config file or env var
-      @@config[:access_key] = ""
+      @@config[:access_key] = ''
 
       # The oboe Ruby client has the ability to sanitize query literals
       # from SQL statements.  By default this is disabled.  Enable to
       # avoid collecting and reporting query literals to TraceView.
       @@config[:sanitize_sql] = false
 
-      if ENV.has_key?('OPENSHIFT_TRACEVIEW_TLYZER_IP')
+      if ENV.key?('OPENSHIFT_TRACEVIEW_TLYZER_IP')
         # We're running on OpenShift
-        @@config[:tracing_mode] = "always"
+        @@config[:tracing_mode] = 'always'
         @@config[:reporter_host] = ENV['OPENSHIFT_TRACEVIEW_TLYZER_IP']
         @@config[:reporter_port] = ENV['OPENSHIFT_TRACEVIEW_TLYZER_PORT']
       else
         # The default configuration
-        @@config[:tracing_mode] = "through"
-        @@config[:reporter_host] = "127.0.0.1"
-        @@config[:reporter_port] = "7831"
+        @@config[:tracing_mode] = 'through'
+        @@config[:reporter_host] = '127.0.0.1'
+        @@config[:reporter_port] = '7831'
       end
 
       @@config[:verbose] = false
@@ -99,36 +102,41 @@ module Oboe
       @@config[key.to_sym] = value
 
       if key == :sampling_rate
-        Oboe.logger.warn "WARNING: :sampling_rate is not a supported setting for Oboe::Config.  Please use :sample_rate."
+        Oboe.logger.warn 'sampling_rate is not a supported setting for Oboe::Config.  ' \
+                         'Please use :sample_rate.'
       end
 
       if key == :sample_rate
-        unless value.is_a?(Integer) or value.is_a?(Float)
-          raise "oboe :sample_rate must be a number between 1 and 1000000 (1m)"
+        unless value.is_a?(Integer) || value.is_a?(Float)
+          fail 'oboe :sample_rate must be a number between 1 and 1000000 (1m)'
         end
 
         # Validate :sample_rate value
         unless value.between?(1, 1e6)
-          raise "oboe :sample_rate must be between 1 and 1000000 (1m)"
+          fail 'oboe :sample_rate must be between 1 and 1000000 (1m)'
         end
 
         # Assure value is an integer
         @@config[key.to_sym] = value.to_i
 
-        Oboe.set_sample_rate(value)
+        Oboe.set_sample_rate(value) if Oboe.loaded
       end
 
       # Update liboboe if updating :tracing_mode
       if key == :tracing_mode
-        Oboe.set_tracing_mode(value)
+        Oboe.set_tracing_mode(value) if Oboe.loaded
       end
+    end
+
+    def self.instrumentation_list
+      @@instrumentation
     end
 
     def self.method_missing(sym, *args)
       if sym.to_s =~ /(.+)=$/
         self[$1] = args.first
       else
-        unless @@config.has_key?(sym)
+        unless @@config.key?(sym)
           Oboe.logger.warn "[oboe/warn] Unknown method call on Oboe::Config: #{sym}"
         end
         self[sym]
@@ -138,4 +146,3 @@ module Oboe
 end
 
 Oboe::Config.initialize
-